*** empty log message ***
git-svn-id: file:///usr/local/svn/fsc2@3541 934aee4a-8510-0410-ae50-ec5b2ea4bb86

+++ trunk/fsc2/src/func_save.c
@@ -590,7 +590,7 @@ Var *f_save( Var *v )
 
 static int print_array( Var *v, long cur_dim, long *start, int fid )
 {
-	long i;
+	unsigned int i;
 	int total_count = 0;
 	int count = 0;
 
@@ -628,7 +628,7 @@ static int print_array( Var *v, long cur_dim, long *start, int fid )
 
 static int print_slice( Var *v, int fid )
 {
-	long i;
+	unsigned int i;
 	int count = 0;
